tests/wpt/tests/shared-storage/select-url-on-two-worklets.tentative.https.sub.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/shared-storage/resources/util.js"></script>
<script src="/fenced-frame/resources/utils.js"></script>

<body>
<script>
'use strict';

promise_test(async () => {
  const ancestor_key = token();
  let url0 = generateURL("/shared-storage/resources/frame0.html",
                         [ancestor_key]);
  let url1 = generateURL("/shared-storage/resources/frame1.html",
                         [ancestor_key]);

  let worklet0 = await sharedStorage.createWorklet("/shared-storage/resources/simple-module.js");
  let worklet1 = await sharedStorage.createWorklet("/shared-storage/resources/simple-module.js");

  let select_url_result0 = await worklet0.selectURL(
      "increment-global-variable-and-return-original-value-operation",
      [{url: url0}, {url: url1}], {resolveToConfig: true, keepAlive: true});
  assert_true(validateSelectURLResult(select_url_result0, true));
  attachFencedFrame(select_url_result0, 'opaque-ads');
  const result0 = await nextValueFromServer(ancestor_key);
  assert_equals(result0, "frame0_loaded");

  let select_url_result1 = await worklet1.selectURL(
      "increment-global-variable-and-return-original-value-operation",
      [{url: url0}, {url: url1}], {resolveToConfig: true, keepAlive: true});
  assert_true(validateSelectURLResult(select_url_result1, true));
  attachFencedFrame(select_url_result1, 'opaque-ads');
  const result1 = await nextValueFromServer(ancestor_key);

  // This indicates that the prior worklet0.selectURL and worklet1.selectURL ran
  // in two different worklet environments, as they did not share the global
  // variable.
  assert_equals(result1, "frame0_loaded");

}, 'selectURL() on two explicitly created SharedStorageWorklet. The two ' +
   'calls should run in two different worklet environments.');

</script>
</body>
